Fix for issue #155: 收款渠道功能，页面上有大量的0.0000000000000000000000000000
Automatically generated by Gemini Bot.

+++ src/Aiursoft.EmployeeCenter.Entities/CollectionChannel.cs
@@ -20,7 +20,7 @@ public class CollectionChannel
     [ForeignKey(nameof(ContractId))]
     public Contract? Contract { get; set; }
 
-    public decimal ReferenceAmount { get; set; }
+    public long ReferenceAmount { get; set; }
 
     [MaxLength(10)]
     public required string Currency { get; set; }
